Разбил на модули файлы .js
